requirements.txt
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
boto3==1.18.23
botocore==1.21.23
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coverage==5.5
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-icons==0.6.4
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-sslserver==0.22
django-storages==1.11.1
django-utils-six==2.0
easy-thumbnails==2.7.1
environ==1.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
oauthlib==3.1.1
pep8==1.7.1
Pillow==8.3.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
sendgrid==6.8.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
